bonding: Incorrect TX queue offset
Patchwork-id: 98894
O-Subject: [RHEL6.7 PATCH BZ 1159818 185/657] bonding: Incorrect TX queue offset
Bugzilla: 1159818
RH-Acked-by: Marcelo Leitner <mleitner@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

From: Phil Oester <kernel@linuxace.com>

When packets come in from a device with >= 16 receive queues
headed out a bonding interface, syslog gets filled with this:

    kernel: bond0 selects TX queue 16, but real number of TX queues is 16

because queue_mapping is offset by 1.  Adjust return value
to account for the offset.

This is a revision of my earlier patch (which did not use the
skb_rx_queue_* helpers - thanks to Ben for the suggestion).
Andy submitted a similar patch which emits a pr_warning on
invalid queue selection, but I believe the log spew is
not useful.  We can revisit that question in the future,
but in the interim I believe fixing the core problem is
worthwhile.

Signed-off-by: Phil Oester <kernel@linuxace.com>
Signed-off-by: Andy Gospodarek <andy@greyhouse.net>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit fd0e435b0fe85622f167b84432552885a4856ac8)

Signed-off-by: Rafael Aquini <aquini@redhat.com>