bonding: fix bond_release_all inconsistencies
Patchwork-id: 99013
O-Subject: [RHEL6.7 PATCH BZ 1159818 300/657] bonding: fix bond_release_all inconsistencies
Bugzilla: 1159818
RH-Acked-by: Marcelo Leitner <mleitner@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

From: "nikolay@redhat.com" <nikolay@redhat.com>

This patch fixes the following inconsistencies in bond_release_all:
- IFF_BONDING flag is not stripped from slaves
- MTU is not restored
- no netdev notifiers are sent
Instead of trying to keep bond_release and bond_release_all in sync
I think we can re-use bond_release as the environment for calling it
is correct (RTNL is held). I have been running tests for the past
week and they came out successful. The only way for bond_release to fail
is for the slave to be attached in a different bond or to not be a slave
but that cannot happen as RTNL is held and no slave manipulations can be
achieved.

V2: As suggested bond_release is renamed to __bond_release_one with a
new parameter "all" introduced so to avoid calling unnecessary code while
destroying a bond, and a wrapper for it called bond_release is created
because of ndo_del_link. bond_release_all() is removed.

Signed-off-by: Nikolay Aleksandrov <nikolay@redhat.com>
Signed-off-by: Jay Vosburgh <fubar@us.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 0896341a44bf04bf6149d9307fe4686006f3eee1)

Conflicts:
	drivers/net/bonding/bond_main.c
	- Minor differences in bond_release_all(). I removed it anyway.

Signed-off-by: Rafael Aquini <aquini@redhat.com>