bonding: rhel-specific: move bond_attach/detach_slave in the proper position
Patchwork-id: 99101
O-Subject: [RHEL6.7 PATCH BZ 1159818 393/657] bonding: RHEL-specific: move bond_attach/detach_slave in the proper position
Bugzilla: 1159818
RH-Acked-by: Marcelo Leitner <mleitner@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

This should move the attach/detach slave in the end of bond_enslave and
the beginning of bond_release_one respectively since that is the where
the slaves get linked in upstream (using neighbor lists though).
This will help us apply a couple of future fixes and patches that depend
on this enslave/release organization.

Signed-off-by: Nikolay Aleksandrov <nikolay@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>