bonding: fix rtnl: assertion failed at net/core/rtnetlink.c for ab arp monitor
Patchwork-id: 99266
O-Subject: [RHEL6.7 PATCH BZ 1159818 563/657] bonding: Fix RTNL: assertion failed at net/core/rtnetlink.c for ab arp monitor
Bugzilla: 1159818
RH-Acked-by: Marcelo Leitner <mleitner@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

From: dingtianhong <dingtianhong@huawei.com>

Veaceslav has reported and fix this problem by commit f2ebd477f141bc0
(bonding: restructure locking of bond_ab_arp_probe()). According Jay's
opinion, the current solution is not very well, because the notification
is to indicate that the interface has actually changed state in a meaningful
way, but these calls in the ab ARP monitor are internal settings of the flags
to allow the ARP monitor to search for a slave to become active when there are
no active slaves. The flag setting to active or backup is to permit the ARP
monitor's response logic to do the right thing when deciding if the test
slave (current_arp_slave) is up or not.

So the best way to fix the problem is that we should not send a notification
when the slave is in testing state, and check the state at the end of the
monitor, if the slave's state recover, avoid to send pointless notification
twice. And RTNL is really a big lock, hold it regardless the slave's state
changed or not when the current_active_slave is null will loss performance
(every 100ms), so we should hold it only when the slave's state changed and
need to notify.

I revert the old commit and add new modifications.

Cc: Jay Vosburgh <fubar@us.ibm.com>
Cc: Veaceslav Falico <vfalico@redhat.com>
Cc: Andy Gospodarek <andy@greyhouse.net>
Signed-off-by: Ding Tianhong <dingtianhong@huawei.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit b0929915e0356acedf59504521c097ecada88b19)

Nik: remove the gfp argument from rtmsg_ifinfo since we don't have it.
Later the locks that were problematic will be removed so commit:
072256d1f2b8ba0bbb265d590c703f3d57a39d6a bonding: make slave status
notifications GFP_ATOMIC

will be unnecessary in the end and we can still use the notifications.

Signed-off-by: Rafael Aquini <aquini@redhat.com>