revert "revert "bonding: Fix pktgen to not oops on unsupported drivers""
Patchwork-id: 99401
O-Subject: [RHEL6.7 PATCH BZ 1159818 658/657] Revert "Revert [netdrv] Fix pktgen to not oops on unsupported drivers"
Bugzilla: 1159818
RH-Acked-by: Marcelo Leitner <mleitner@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

This reverts commit d1801731d76bc61df7a0011565832e2681077a67.
We need this fix to prevent the oops in the bonding, this is only
the bonding part which was reverted. Fortunately the fix is small and
straightforward.

Brew with this patch - https://brewweb.devel.redhat.com/taskinfo?taskID=8195298

Signed-off-by: Nikolay Aleksandrov <nikolay@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>