nfs: Make sure pre_change_attr is initialized correctly
Patchwork-id: 99406
O-Subject: [RHEL6 PATCH] [bz1160042] nfs: Make sure pre_change_attr is initialized correctly
Bugzilla: 1160042
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Fixes bug 1160042.

When upstream commit 3a1556e8 (NFSv2/v3: Simulate the change attribute)
was backported to RHEL 6 for bug 1018653, the part that initializes the
pre-op change attribute was missed.  As a result, we can wind up with a
lot of unnecessary data cache invalidations on NFS v3 and a lot more
on-the-wire NFS READ operations as a result.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>