Fix ring_buffer perf_output_space() boundary calculation
Patchwork-id: 99456
O-Subject: [PATCH RHEL6.7 BZ1159929 019/320] perf: Fix ring_buffer perf_output_space() boundary calculation
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit dd9c086d9f507d02d5ba4d7c5eef4bb9518088b8
Author: Stephane Eranian <eranian@google.com>
Date: Mon Mar 18 14:33:28 2013 +0100

description
===========
This patch fixes a flaw in perf_output_space(). In case the size
of the space needed is bigger than the actual buffer size, there
may be situations where the function would return true (i.e.,
there is space) when it should not. head > offset due to
rounding of the masking logic.

The problem can be tested by activating BTS on Intel processors.
A BTS record can be as big as 16 pages. The following command
fails:

  $ perf record -m 4 -c 1 -e branches:u my_test_program

You will get a buffer corruption with this. Perf report won't be
able to parse the perf.data.

The fix is to first check that the requested space is smaller
than the buffer size. If so, then the masking logic will work
fine. If not, then there is no chance the record can be saved
and it will be gracefully handled by upper code layers.

[ In v2, we also make the logic for the writable more explicit by
  renaming it to rb->overwrite because it tells whether or not the
  buffer can overwrite its tail (suggested by PeterZ). ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>