x86: Fix constraint table end marker bug
Patchwork-id: 99471
O-Subject: [PATCH RHEL6.7 BZ1159929 028/320] perf/x86: Fix constraint table end marker bug
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit cf30d52e2d11523c42048ab89ed4215b5021526a
Author: Maria Dimakopoulou <maria.n.dimakopoulou@gmail.com>
Date: Thu Dec 5 01:24:37 2013 +0200

description
===========
The EVENT_CONSTRAINT_END() macro defines the end marker as
a constraint with a weight of zero. This was all fine
until we blacklisted the corrupting memory events on
Intel IvyBridge. These events are blacklisted by using
a counter bitmask of zero. Thus, they also get a constraint
weight of zero.

The iteration macro: for_each_constraint tests the weight==0.
Therefore, it was stopping at the first blacklisted event, i.e.,
0xd0. The corrupting events were therefore considered as
unconstrained and were scheduled on any of the generic counters.

This patch fixes the end marker to have a weight of -1. With
this, the blacklisted events get an empty constraint and cannot
be scheduled which is what we want for now.

Signed-off-by: Rafael Aquini <aquini@redhat.com>