trace: Use vfs_getname hook if available
Patchwork-id: 99617
O-Subject: [PATCH RHEL6.7 BZ1159929 193/320] perf trace: Use vfs_getname hook if available
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit c522739d72a341a3e74a369ce6298b9412813d3f
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Sep 27 18:06:19 2013 -0300

description
===========
Initially it tries to find a probe:vfs_getname that should be setup
with:

 perf probe 'vfs_getname=getname_flags:65 pathname=result->name:string'

or with slight changes to cope with code flux in the getname_flags code.

In the future, if a "vfs:getname" tracepoint becomes available, then it
will be preferred.

This is not strictly required and more expensive method of reading the
/proc/pid/fd/ symlink will be used when the fd->path array entry is not
populated by a previous vfs_getname + open syscall ret sequence.

As with any other 'perf probe' probe the setup must be done just once
and the probe will be left inactive, waiting for users, be it 'perf
trace' of any other tool.

Signed-off-by: Rafael Aquini <aquini@redhat.com>