list: Show error if tracepoints not available
Patchwork-id: 99637
O-Subject: [PATCH RHEL6.7 BZ1159929 209/320] perf list: Show error if tracepoints not available
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit f11cfc6f294dbd83b0d58037404df2bd16066238
Author: Pekka Enberg <penberg@kernel.org>
Date: Tue Oct 15 23:07:27 2013 +0300

description
===========
Tracepoints are not visible in "perf list" on Fedora 19 because regular
users have no permission to /sys/kernel/debug by default. Show an error
message so that the user knows about it instead of assuming that
tracepoints are not supported on the system.

Signed-off-by: Rafael Aquini <aquini@redhat.com>