top: Add --max-stack option to limit callchain stack scan
Patchwork-id: 99700
O-Subject: [PATCH RHEL6.7 BZ1159929 216/320] perf top: Add --max-stack option to limit callchain stack scan
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit 5dbb6e81d85e55ee2b4cf523c1738e16f63e5400
Author: Waiman Long <Waiman.Long@hp.com>
Date: Fri Oct 18 10:38:49 2013 -0400

description
===========
When the callgraph function is enabled (-G), it may take a long time to
scan all the stack data and merge them accordingly.

This patch adds a new --max-stack option to perf-top to limit the depth
of callchain stack data to look at to reduce the time it takes for
perf-top to finish its processing. It reduces the amount of information
provided to the user in exchange for faster speed.

Signed-off-by: Rafael Aquini <aquini@redhat.com>