tools: Use an accessor to read thread comm
Patchwork-id: 99655
O-Subject: [PATCH RHEL6.7 BZ1159929 235/320] perf tools: Use an accessor to read thread comm
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit b9c5143a012a543c4ee872498d6dbae5c10beb2e
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date: Wed Sep 11 14:46:56 2013 +0200

description
===========
As the thread comm is going to be implemented by way of a more
complicated data structure than just a pointer to a string from the
thread struct, convert the readers of comm to use an accessor instead of
accessing it directly.

The accessor will be later overriden to support an enhanced comm
implementation.

[ Rename thread__comm_curr() to thread__comm_str() ]
[ Fixed up some minor const pointer issues ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>