traceevent: Update printk formats when entered
Patchwork-id: 99666
O-Subject: [PATCH RHEL6.7 BZ1159929 252/320] tools lib traceevent: Update printk formats when entered
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit 18900af8292180151c82f0762506fa0740aa54a5
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Fri Nov 1 17:53:54 2013 -0400

description
===========
Instead of cropping off the '"' and '\n"' from a printk format every
time it is referenced, do it when it's added. This makes it easier to
reference a printk_map and should speed things up a little.

Signed-off-by: Rafael Aquini <aquini@redhat.com>