traceevent: Check for spaces in character array
Patchwork-id: 99703
O-Subject: [PATCH RHEL6.7 BZ1159929 256/320] tools lib traceevent: Check for spaces in character array
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit 5efb9fbd5f1bfe4435bd0a3ea5f0e187875509c2
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Fri Nov 1 17:53:58 2013 -0400

description
===========
Currently when using the raw format for fields, when looking at a
character array, to determine if it is a string or not, we make sure all
characters are "isprint()". If not, then we consider it a numeric array,
and print the hex numbers of the characters instead.

But it seems that '\n' fails the isprint() check! Add isspace() to the
check as well, such that if all characters pass isprint() or isspace()
it will assume the character array is a string.

Signed-off-by: Rafael Aquini <aquini@redhat.com>