traceevent: Add flags NOHANDLE and PRINTRAW to individual events
Patchwork-id: 99706
O-Subject: [PATCH RHEL6.7 BZ1159929 257/320] tools lib traceevent: Add flags NOHANDLE and PRINTRAW to individual events
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit c6c2b960b7a4105f096499fba3df65d6c0272a20
Author: Steven Rostedt <srostedt@redhat.com>
Date: Fri Nov 1 17:53:59 2013 -0400

description
===========
Add the flags EVENT_FL_NOHANDLE and EVENT_FL_PRINTRAW to the event flags
to have the event either ignore the register handler or to ignore the
handler and also print the raw format respectively.

This allows a tool to force a raw format or non handle for an event.

Signed-off-by: Rafael Aquini <aquini@redhat.com>