fs: Rename NAME_find_mountpoint() to NAME__mountpoint()
Patchwork-id: 99681
O-Subject: [PATCH RHEL6.7 BZ1159929 265/320] perf fs: Rename NAME_find_mountpoint() to NAME__mountpoint()
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit cf38fadade52df937521dd70d4437df1a9354cd9
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Nov 5 14:48:50 2013 -0300

description
===========
Shorten it, "finding" it is an implementation detail, what callers want
is the pathname, not to ask for it to _always_ do the lookup.

And the existing implementation already caches it, i.e. it doesn't
"finds" it on every call.

Signed-off-by: Rafael Aquini <aquini@redhat.com>