trace: Beautify fifth argument of mmap() as fd
Patchwork-id: 99710
O-Subject: [PATCH RHEL6.7 BZ1159929 286/320] perf trace: Beautify fifth argument of mmap() as fd
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit 73faab3a421351c342b10b0177fec9eea1a1d809
Author: Namhyung Kim <namhyung.kim@lge.com>
Date: Tue Nov 12 15:24:59 2013 +0900

description
===========
The fifth argument of mmap syscall is fd and it often contains -1 as a
value for anon mappings.  Without this patch it doesn't show the file
name as well as it shows -1 as 4294967295.

Signed-off-by: Rafael Aquini <aquini@redhat.com>