tools: Synthesize anon MMAP records again
Patchwork-id: 99740
O-Subject: [PATCH RHEL6.7 BZ1159929 307/320] perf tools: Synthesize anon MMAP records again
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit 9d4ecc8893832337daf241236841db966fa53489
Author: Don Zickus <dzickus@redhat.com>
Date: Wed Nov 13 15:32:06 2013 -0300

description
===========
When introducing the PERF_RECORD_MMAP2 in:

5c5e854bc760 perf tools: Add attr->mmap2 support

A check for the number of entries parsed by sscanf was introduced that
assumed all of the 8 fields needed to be correctly parsed so that
particular /proc/pid/maps line would be considered synthesizable.

That broke anon records synthesizing, as it doesn't have the 'execname'
field.

Fix it by keeping the sscanf return check, changing it to not require
that the 'execname' variable be parsed, so that the preexisting logic
can kick in and set it to '//anon'.

This should get things like JIT profiling working again.

[ commit log message is mine, dzickus reported the problem with a patch ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>