traceevent: Fix use of multiple options in processing field
Patchwork-id: 99754
O-Subject: [PATCH RHEL6.7 BZ1159929 317/320] tools lib traceevent: Fix use of multiple options in processing field
Bugzilla: 1159929
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1159929
https://bugzilla.redhat.com/show_bug.cgi?id=1159929

upstream
========
commit eff2c92f86c2ac2a0eab3749d58be39592293c3a
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Mon Nov 18 14:23:14 2013 -0500

description
===========
Jiri Olsa reported that the scsi_dispatch_cmd_done event failed to parse
with:

  Error: expected type 5 but read 4
  Error: expected type 5 but read 4

The problem is with this part of the print_fmt:

  __print_symbolic(((REC->result) >> 24) & 0xff, ...

The __print_symbolic() helper function's first parameter is the field to
use to determine what symbol to print based on the value of the result.
The parser can handle one operation, but it can not handle multiple
operations ('>>' and '&').

Add code to process all operations for the field argument for
__print_symbolic() as well as __print_flags().

Signed-off-by: Rafael Aquini <aquini@redhat.com>