Revert "watchdog: quiet down the boot messages"
Patchwork-id: 99891
O-Subject: [RHEL6 PATCH 1/5] Revert "[watchdog] quiet down the boot messages"
Bugzilla: 1111822
RH-Acked-by: Ulrich Obergfell <uobergfe@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1111822

This reverts commit a21807eb17497de6a0657162f8f11df138c97867.

Replace the RHEL version of this patch with the upstream version.
This patch reverts the RHEL version, the next patch backports
the upstream version. This converts a local variable to a static
global, which is necessary for the next patch, as well as brings
in some context that helps us backport the remaining patches in
this series.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>