ring-buffer: Check if ring buffer iterator has stale data
Patchwork-id: 99952
O-Subject: [PATCH RHEL6.7 BZ1123102 1/2] ring-buffer: Check if ring buffer iterator has stale data
Bugzilla: 1123102
RH-Acked-by: Steven Rostedt <srostedt@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1123102
https://bugzilla.redhat.com/show_bug.cgi?id=1123102

upstream
========
commit 492a74f4210e15f4701422e2e1c4cd3c1e45ddae
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Mon Jan 25 15:17:47 2010 -0500

description
===========
Usually reads of the ring buffer is performed by a single task.
There are two types of reads from the ring buffer.

One is a consuming read which will consume the entry that was read
and the next read will be the entry that follows.

The other is an iterator that will let the user read the contents of
the ring buffer without modifying it. When an iterator is allocated,
writes to the ring buffer are disabled to protect the iterator.

The problem exists when consuming reads happen while an iterator is
allocated. Specifically, the kind of read that swaps out an entire
page (used by splice) and replaces it with a new read. If the iterator
is on the page that is swapped out, then the next read may read
from this swapped out page and return garbage.

This patch adds a check when reading the iterator to make sure that
the iterator contents are still valid. If a consuming read has taken
place, the iterator is reset.

Cc: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Rafael Aquini <aquini@redhat.com>