cifs: always update the inode cache with the results from a FIND_*
Patchwork-id: 100243
O-Subject: [RHEL6 PATCH fs] [bz1163010] cifs: always update the inode cache with the results from a FIND_*
Bugzilla: 1163010
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

From: Jeff Layton <jlayton@redhat.com>

Fixes bz 1163010

We can end up with multiple files sharing the same inode number on a
cifs share. The reproducer which leads to this condition is provided
in the bz. The patch was successfully tested against the reproduce.

Original summary follows
--

When we get back a FIND_FIRST/NEXT result, we have some info about the
dentry that we use to instantiate a new inode. We were ignoring and
discarding that info when we had an existing dentry in the cache.

Fix this by updating the inode in place when we find an existing dentry
and the uniqueid is the same.

Cc: <stable@vger.kernel.org> # .31.x
Reported-and-Tested-by: Andrew Bartlett <abartlet@samba.org>
Reported-by: Bill Robertson <bill_robertson@debortoli.com.au>
Reported-by: Dion Edwards <dion_edwards@debortoli.com.au>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit cd60042cc1392e79410dc8de9e9c1abb38a29e57)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Reported-by:  HONDA Hirofumi <honda.hirofumi@nttcom.co.jp>
Signed-off-by: Rafael Aquini <aquini@redhat.com>