nfsd: disable 4.1 in rhel6
Patchwork-id: 100268
O-Subject: [Patch RHEL6 fs] nfsd: disable 4.1 in rhel6
Bugzilla: 1122493
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Jesper Brouer <jbrouer@redhat.com>
RH-Acked-by: Sachin Prabhu <sprabhu@redhat.com>

NFSv4.1 in RHEL6 has a number of bugs and this is a distraction for QE.
The most prominent user-visible 4.1 feature is pNFS, which we haven't
implemented.  So it's unlikely anyone will put up a big fuss about lack
of 4.1 support in RHEL6.

In Fedora, for example, we wanted 4.1 to be easily available for testers
and early adopters, so it was available but off by default.  In RHEL6 to
minimize the chance of having to support it we'd rather it just be off
completely.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Bugzilla: 1122493
Signed-off-by: Rafael Aquini <aquini@redhat.com>