fuse: prevent null nd panic on dentry revalidate
Patchwork-id: 100827
O-Subject: [RHEL6 PATCH] fuse: prevent null nd panic on dentry revalidate
Bugzilla: 1162782
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

This is a rhel6-specific patch to address a regression introduced in the
following rhel6 commit:

	[fs] fuse: ignore entry-timeout LOOKUP_REVAL

This commit checks nd->flags, but nd is not always passed to
d_revalidate(). If not, this leads to a NULL dereference and panic. This
was not obvious at the time of the original backport as nd was not
passed to the upstream variant of d_revalidate() at that point in time.

Update fuse_dentry_revalidate() to check that nd is not NULL before it
is dereferenced.

Signed-off-by: Brian Foster <bfoster@redhat.com>