memcg: fix crash in re-entrant cgroup_clear_css_refs()
Patchwork-id: 101619
O-Subject: [PATCH RHEL6] [mm] memcg: fix crash in re-entrant cgroup_clear_css_refs()
Bugzilla: 1168185
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

Bugzilla: 1168185

cgroup_clear_css_refs() BUG_ON-checks that the refcounts encountered
are not 0 already, to catch erroneous code flow somewhere else.  But
that function is entered repeatedly after the following fix specific
to RHEL6:

commit f3a7842b6e2b46024caec94e4d14f14164ec0c5f
Author: Johannes Weiner <jweiner@redhat.com>
Date:   Fri Jun 20 21:12:13 2014 -0400

    [mm] memcg: fix race condition between memcg teardown and swapin

    Message-id: <20140620211213.GJ7331@redhat.com>
    Patchwork-id: 84302
    O-Subject: [PATCH RHEL6.6 BZ1001197 V3] mm: memcg: fix race condition between memcg teardown and swapin
    Bugzilla: 1001197
    RH-Acked-by: Rafael Aquini <aquini@redhat.com>
    RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

It's not enough to remove the BUG_ON(), because once the css refs are
cleared and trygets can fail, we really don't want to fail the removal
and leave this zombie cgroup dangling visible to the user.

Bite the bullet and backport the no-fail reparenting that is already
done upstream: rather than charging the parent and uncharging the
child, acknowledge that the parent is already charged hierarchically,
so it's enough to cancel the charge in the child.  Once that is done
and reparenting can't fail, simply do it from .destroy() when the css
refs are cleared and no new references can show up.

Signed-off-by: Johannes Weiner <jweiner@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>