KEYS: Add invalidation support
Patchwork-id: 101764
O-Subject: [RHEL6 PATCH fs 1/2] [bz885899] KEYS: Add invalidation support
Bugzilla: 885899
RH-Acked-by: David Howells <dhowells@redhat.com>

From: David Howells <dhowells@redhat.com>

Add support for invalidating a key - which renders it immediately invisible to
further searches and causes the garbage collector to immediately wake up,
remove it from keyrings and then destroy it when it's no longer referenced.

It's better not to do this with keyctl_revoke() as that marks the key to start
returning -EKEYREVOKED to searches when what is actually desired is to have the
key refetched.

To invalidate a key the caller must be granted SEARCH permission by the key.
This may be too strict.  It may be better to also permit invalidation if the
caller has any of READ, WRITE or SETATTR permission.

The primary use for this is to evict keys that are cached in special keyrings,
such as the DNS resolver or an ID mapper.

Signed-off-by: David Howells <dhowells@redhat.com>

The patch has been tested using the cifs module doing krb5
authentication as part of the reproducer in bz885899.

Signed-off-by: Sachin Prabhu<sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>