cifs: Use key_invalidate instead of the rh_key_invalidate()
Patchwork-id: 101765
O-Subject: [RHEL6 PATCH fs 2/2] [bz885899]cifs: Use key_invalidate instead of the rh_key_invalidate()
Bugzilla: 885899
RH-Acked-by: Niels de Vos <ndevos@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

rh_key_invalidate() was used as a workaround instead of key_invalidate()
which didn't exist for RHEL 6. This however doesn't seem to call the gc
correctly resulting in keyrings full of revoked keys.

To address this problem, we have decided to bite the bullet and backport
the original key_invalidate() in the earlier patch. This patch removes
the calls to rh_key_invalidate and replaces them with key_invalidate()
instead.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>