timechart: Group figures and add title with details
Patchwork-id: 101816
O-Subject: [PATCH RHEL6.7 BZ1172076 026/313] perf timechart: Group figures and add title with details
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit cbb2e81e5232b1bca5cd2aa1d7a7eb1cd30f8304
Author: Stanislav Fomichev <stfomichev@yandex-team.ru>
Date: Fri Nov 1 20:25:49 2013 +0400

description
===========
Add titles to figures so we can run SVG interactively in Firefox and
check event details in the tooltips.

This also aids exploring SVG with Inkscape because when user clicks on
one part of logical figure, all parts are selected.

It's also possible to read titles with Inkscape in the object details.

Signed-off-by: Rafael Aquini <aquini@redhat.com>